+++ src/controllers/CompressController.php
@@ -10,6 +10,7 @@
 
 namespace venveo\compress\controllers;
 
+use craft\helpers\App;
 use craft\helpers\DateTimeHelper;
 use craft\web\Controller;
 use venveo\compress\Compress as Plugin;
@@ -17,6 +18,7 @@ use venveo\compress\errors\CompressException;
 use venveo\compress\models\Archive as ArchiveModel;
 use venveo\compress\records\Archive as ArchiveRecord;
 use yii\base\InvalidConfigException;
+use yii\web\RangeNotSatisfiableHttpException;
 
 /**
  * Class CompressController
@@ -32,6 +34,7 @@ class CompressController extends Controller
      * @return \yii\web\Response
      * @throws CompressException
      * @throws InvalidConfigException
+     * @throws RangeNotSatisfiableHttpException
      */
     public function actionGetLink($uid)
     {
@@ -41,19 +44,15 @@ class CompressController extends Controller
             return \Craft::$app->response->setStatusCode(404, 'Archive could not be found');
         }
 
-        // If the asset is ready, redirect to its URL
+        // If the asset is ready, send it over
         if ($record->assetId) {
             $archiveModel = ArchiveModel::hydrateFromRecord($record);
             // It's possible for an asset ID to exist, but getAsset to return false on soft-deleted assets
-            if ($archiveModel->getAsset()) {
-                $assetUrl = $archiveModel->getAsset()->getUrl();
-                if ($assetUrl) {
-                    $record->dateLastAccessed = DateTimeHelper::currentUTCDateTime();
-                    $record->save();
-                    return \Craft::$app->response->redirect($archiveModel->getAsset()->getUrl());
-                }
-
-                return \Craft::$app->response->setStatusCode(404, 'Could not produce zip file URL.');
+            $archiveAsset = $archiveModel->getAsset();
+            if ($archiveAsset) {
+                $record->dateLastAccessed = DateTimeHelper::currentUTCDateTime();
+                $record->save();
+                return $this->getAssetResponse($archiveModel->getAsset());
             }
         }
 
@@ -71,11 +70,35 @@ class CompressController extends Controller
                     \Craft::$app->cache->delete($cacheKey . ':jobId');
                 }
             }
-            return \Craft::$app->response->redirect($archiveModel->asset->getUrl());
+            return $this->getAssetResponse($archiveModel->getAsset());
         } catch (\Exception $e) {
             \Craft::error('Archive could not be generated: ' . $e->getMessage(), __METHOD__);
             \Craft::error($e->getTraceAsString(), __METHOD__);
-            throw new CompressException('Archive could not be generated: ' . $e->getMessage());
+            return \Craft::$app->response->setStatusCode(404, 'Could not produce zip file URL.');
         }
+    }
+
+    /**
+     * @param $archiveAsset
+     * @return \craft\web\Response|\yii\console\Response|\yii\web\Response
+     * @throws RangeNotSatisfiableHttpException
+     */
+    protected function getAssetResponse($archiveAsset) {
+        if (!$archiveAsset) {
+            return \Craft::$app->response->setStatusCode(404, 'Could not produce zip file URL.');
+        }
+        $assetUrl = $archiveAsset->getUrl();
+
+        // If we have a public URL for the asset, we'll just 302 redirect to it
+        if ($assetUrl) {
+            return \Craft::$app->response->redirect($archiveAsset->getUrl());
+        }
+        App::maxPowerCaptain();
+        // No public URLs, we'll need to stream the response.
+        return $this->response
+            ->sendStreamAsFile($archiveAsset->getStream(), $archiveAsset->getFilename(), [
+                'fileSize' => $archiveAsset->size,
+                'mimeType' => $archiveAsset->getMimeType(),
+            ]);
     }
 }